[nibeheatpump] #5012, #5604 Channel update logic, to and from modbus data conversion fixes (#5015)
* Update channels even if value did not change
* NibeGW to support ENC28J60
* Less memory footprint
* Read UDP packet flushing was missing
* Do periodic updates to channels even if value did not change
* Negative signed 8bit values are not parsed correctly and caching check fix
* UDP Flush is only required for TRANSPORT_ETH_ENC28J60
* Negative signed values convert fixes + tests
* Proper number convert from Modbus
* BigDecimal convert to Modbus fixes. Tests rewrite
* Command type checking
* MODBUS_READ_RESPONSE_MSG is now converted properly regardless of Item type
* More command types support
* Test improvements
* READ_OUT message values are read twice in refreshInterval period if not changed
* Some NibeDataType.S32 are actually NibeDataType.U32
* TRANSPORT_ETH_ENC28J60 moved closer to PRODINO_BOARD
* Credits to repo with int convert methods
* Simpler method to convert. Optimize imports and reformat
* Removed unnecessary cast, had to change values in tests to signed
* More tests for edge cases of value converts
* EthernetUdp import moved

Fixes #5012
Fixes #5604

+++ bundles/org.openhab.binding.nibeheatpump/src/main/java/org/openhab/binding/nibeheatpump/internal/message/ModbusDataReadOutMessage.java
@@ -101,7 +101,7 @@ public class ModbusDataReadOutMessage extends NibeHeatPumpBaseMessage {
                 for (int i = NibeHeatPumpProtocol.OFFSET_DATA; i < (msglen - 1); i += 4) {
 
                     int id = ((rawMessage[i + 1] & 0xFF) << 8 | (rawMessage[i + 0] & 0xFF));
-                    int value = (short) ((rawMessage[i + 3] & 0xFF) << 8 | (rawMessage[i + 2] & 0xFF));
+                    int value = (rawMessage[i + 3] & 0xFF) << 8 | (rawMessage[i + 2] & 0xFF);
 
                     if (id != 0xFFFF) {
                         vals.add(new ModbusValue(id, value));
